Add proper structure
Trying a simple GET request with basicAuth.

Username and Password newtypes are copied from the authHook of
bloodhound, except I've changed them to ByteString instead of Text for
the sake of applyBasicAuth.

Going through bloodhound/bazaari at the same time to wrap my head around
how it's supposed to be structured.