Restructure tests
I've made a helper to remove some duplication in the test code.  It
seems that there isn't a simpler way to get a ReceivingId for now.

At first I tried to cancel each receiving after a test, but that
doesn't actually remove that receiving, it just marks it as "cancelled"
in your shipwire account. So it's not exactly a proper "teardown" step
in the test suite. I ended up removing that.

I've also tried running these tests with a brand new sandbox Shipwire
account. And while it will eventually be possible to programmatically
create products and run tests, I don't think it's possible to create a
Warehouse in any other way than through their
website'sinterface. Meaning we'll have to describe the necessary steps
for people who would want to run the test suite locally or else they
will fail. For example: creating a receiving depends on Warehouse's
region, which gets created based on the name your specify, "My Testing
Warehouse" gets a "MY TE1" region. Not sure if this is a big deal or
not.