Add GET /products endpoint
This adds the necessary functioniality to get back a list of
products.

For some reason the documentation for this endpoint is more
sparse than before. There is no model schema so certain
JSON objects are impossible to understand.

It was also not specified which fields might return "null",
I've had to test and see manually.