Add POST /products for BaseProduct
This allows us to create a `baseProduct`. Still need to implement
`marketingInsert`, `virtualKit` and `kit`.

There are also a few types that got renamed here, the reason for that is
to simplify creating the product. Making type aliases for types that the
library's user will have to use is confusing. For example, it's not
straightforward to make the user use `ValuesResource` instead of
`Values`. It seems that type aliases are more suitable for datatypes we
get back as a response.

I've also started making the test suite independent of the Shipwire
account. It works now, but there are a few more things I need to add.
Specifically, there will be a few steps you'd need to make through the
Shipwire's UI before you can run the tests locally. I've started a draft
of a README with those steps in a different branch.