Add GET /products/{id}
This commit includes a lot of changes to the tests file.

These were done in an attempt to prevent the specs from failing when
you run an individual spec with `:main -m "name"` followed by a full
test suite run.

In order to do that, I had to add the steps of creating a product and
creating a receiving to almost each spec. This has considerably
increased the time it takes to run the whole test suite.

However, if we don't do this, then running a one-off spec might
pollute the Shipwire account, which would later cause the other specs
to fail.

In general, the requirements are something like this: we have to
create a dummy product in order to create a dummy receiving. If a
product has receivings associated with it, we won't be able to
delete (retire) it from the catalog. So it's better to cancel the
receivings ASAP.