Bump dependencies
There was a bug that randomly caused `TlsExceptionHostPort` in older versions
of http-client-tls. It would only happen once in a while and when the tests
were slower, it wouldn't normally happen. Now that the tests are fast, it's
appearing more often.

I fixed it by upgrading the dependencies, as well as moving to a newer resolver
version.